Update config location ; add compose yaml and local image build make target and docs
Update config load to look for a config file in the current working
directory, as a configmap would be placed in production.

Add a make target for local image builds with mimicing the boilerplate
build, but usinng localhost friendly names for use with compose.

Add a compose yaml for easy container development.

Add documentation for the compose and image build setup to
DEVELOPMENT.md

Signed-off-by: Chris Collins <collins.christopher@gmail.com>