adds flag to disable LDAP lookups
LDAP lookups may not be necessary, so this adds a flag to disable the lookups, and instead attempts to send the compliance ticket to the splunk user of the same name.  This may or may not require additional tweaking in the future, though more widespread use of the "REASON" field should remove the need for lookups entirely.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>